feat: Add main components for project collaboration platform
- Implemented Cta component for call-to-action section on home page.
- Developed Features component to showcase platform features.
- Created Hero component for the main introduction section.
- Added HowItWorks component to explain the platform's functionality.
- Built PageFooter component for footer information.
- Designed ProjectShowcase component to display recent projects.
- Introduced ChatDialog component for messaging recruiters.
- Created ProjectCard component to represent individual projects.
- Developed ProjectList component for displaying lists of projects.
- Implemented ProjectsHeader component for search and filter functionality.
- Added SidebarFilters component for skill-based filtering.
- Created SynergizeDialog component for applying to projects.
- Populated mock project data in data.ts for testing.
- Defined Project and ApplicationData types in types/index.ts for type safety.
