+++ components/pages/projects/projects-header.tsx
@@ -1,0 +1,71 @@
+"use client";
+
+import { Input } from "@/components/ui/input";
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
+import { Search } from "lucide-react";
+
+interface ProjectsHeaderProps {
+  searchQuery: string;
+  setSearchQuery: (value: string) => void;
+  selectedType: string;
+  setSelectedType: (value: string) => void;
+  selectedLocation: string;
+  setSelectedLocation: (value: string) => void;
+  locationOptions: string[];
+}
+
+export function ProjectsHeader({
+  searchQuery,
+  setSearchQuery,
+  selectedType,
+  setSelectedType,
+  selectedLocation,
+  setSelectedLocation,
+  locationOptions,
+}: ProjectsHeaderProps) {
+  return (
+    <div className="bg-white border-b sticky top-0 z-40 shadow-sm">
+      <div className="container mx-auto px-4 py-4">
+        <div className="flex flex-col lg:flex-row gap-4">
+          <div className="relative flex-1">
+            <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400 h-5 w-5" />
+            <Input
+              placeholder="Cari proyek, skill, atau kata kunci..."
+              className="pl-10 h-12 text-base"
+              value={searchQuery}
+              onChange={(e) => setSearchQuery(e.target.value)}
+            />
+          </div>
+          <div className="flex gap-2">
+            <Select value={selectedType} onValueChange={setSelectedType}>
+              <SelectTrigger className="w-full sm:w-[140px] h-12">
+                <SelectValue placeholder="Semua Tipe" />
+              </SelectTrigger>
+              <SelectContent>
+                <SelectItem value="all">Semua Tipe</SelectItem>
+                <SelectItem value="Tugas Akhir">Tugas Akhir</SelectItem>
+                <SelectItem value="Lomba">Lomba</SelectItem>
+                <SelectItem value="Riset Dosen">Riset Dosen</SelectItem>
+                <SelectItem value="Startup">Startup</SelectItem>
+                <SelectItem value="Kuliah">Kuliah</SelectItem>
+              </SelectContent>
+            </Select>
+            <Select value={selectedLocation} onValueChange={setSelectedLocation}>
+              <SelectTrigger className="w-full sm:w-[140px] h-12">
+                <SelectValue placeholder="Semua Lokasi" />
+              </SelectTrigger>
+              <SelectContent>
+                <SelectItem value="all">Semua Lokasi</SelectItem>
+                {locationOptions.map((location) => (
+                  <SelectItem key={location} value={location}>
+                    {location}
+                  </SelectItem>
+                ))}
+              </SelectContent>
+            </Select>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
