+++ lib/data.ts
@@ -1,0 +1,163 @@
+export const projects = [
+    {
+      id: 1,
+      title: "Aplikasi Smart Campus IoT",
+      description:
+        "Mengembangkan sistem IoT untuk monitoring fasilitas kampus dengan dashboard real-time dan notifikasi otomatis untuk maintenance",
+      recruiter: {
+        name: "Ahmad Maulana",
+        avatar: "https://img.freepik.com/free-psd/portrait-man-teenager-isolated_23-2151745771.jpg?semt=ais_hybrid&w=740",
+        major: "Teknik Informatika",
+        university: "ITB",
+        rating: 4.8,
+        projects: 12,
+        connections: 150,
+      },
+      skills: ["React", "IoT", "Python", "Arduino", "Node.js"],
+      duration: "3 bulan",
+      members: "2/5",
+      type: "Tugas Akhir",
+      location: "Bandung",
+      posted: "2 hari yang lalu",
+      image: "https://assets-a1.kompasiana.com/items/album/2023/11/14/img-2721-6552c2dbedff761aa670e123.png",
+      urgent: true,
+      views: 45,
+      applicants: 8,
+      deadline: "15 Feb 2024",
+    },
+    {
+      id: 2,
+      title: "Kampanye Sosial Lingkungan",
+      description:
+        "Merancang kampanye kreatif untuk meningkatkan kesadaran lingkungan di kalangan mahasiswa dengan pendekatan multimedia dan storytelling",
+      recruiter: {
+        name: "Sari Rahayu",
+        avatar: "https://mahasiswa.dinus.ac.id/images/foto/A/A11/2022/A11.2022.14148.jpg",
+        major: "Desain Komunikasi Visual",
+        university: "UNPAD",
+        rating: 4.9,
+        projects: 8,
+        connections: 200,
+      },
+      skills: ["Design", "Marketing", "Video Editing", "Social Media", "Adobe Creative"],
+      duration: "2 bulan",
+      members: "4/6",
+      type: "Lomba",
+      location: "Bandung",
+      posted: "1 hari yang lalu",
+      image: "https://relungindonesia.org/wp-content/uploads/2024/09/WhatsApp-Image-2024-03-05-at-16.24.27.jpg",
+      views: 67,
+      applicants: 12,
+      deadline: "20 Feb 2024",
+    },
+    {
+      id: 3,
+      title: "Riset AI untuk Kesehatan",
+      description:
+        "Penelitian pengembangan model AI untuk diagnosis dini penyakit berbasis data medis menggunakan deep learning dan computer vision",
+      recruiter: {
+        name: "Dr. Pratama",
+        avatar: "https://mahasiswa.dinus.ac.id/images/foto/A/A11/2022/A11.2022.14148.jpg",
+        major: "Dosen Peneliti",
+        university: "UI",
+        rating: 5.0,
+        projects: 25,
+        connections: 500,
+      },
+      skills: ["Machine Learning", "Python", "TensorFlow", "Research", "Data Science"],
+      duration: "6 bulan",
+      members: "1/3",
+      type: "Riset Dosen",
+      location: "Jakarta",
+      posted: "3 hari yang lalu",
+      image: "https://itbrief.asia/uploads/story/2024/12/11/techday_26169f03bf0a929134f3.webp",
+      views: 89,
+      applicants: 15,
+      deadline: "28 Feb 2024",
+    },
+    {
+      id: 4,
+      title: "E-commerce Platform UMKM",
+      description:
+        "Membangun platform e-commerce khusus untuk UMKM dengan fitur inventory management, payment gateway, dan analytics dashboard",
+      recruiter: {
+        name: "Budi Santoso",
+        avatar: "https://mahasiswa.dinus.ac.id/images/foto/A/A11/2022/A11.2022.14148.jpg",
+        major: "Sistem Informasi",
+        university: "UGM",
+        rating: 4.7,
+        projects: 15,
+        connections: 180,
+      },
+      skills: ["Laravel", "Vue.js", "MySQL", "Payment Gateway", "AWS"],
+      duration: "4 bulan",
+      members: "3/5",
+      type: "Startup",
+      location: "Yogyakarta",
+      posted: "1 minggu yang lalu",
+      image: "https://bankjombang.co.id/wp-content/uploads/2023/05/tokopedia-platform-e-commerce-yang-paling-dipercaya-umkm-indonesia-jombang-kita.jpg",
+      views: 123,
+      urgent: true,
+      applicants: 20,
+      deadline: "10 Mar 2024",
+    },
+    {
+      id: 5,
+      title: "Game Edukasi Matematika",
+      description:
+        "Mengembangkan game mobile edukatif untuk pembelajaran matematika tingkat SD dengan gamifikasi yang menarik dan sistem reward",
+      recruiter: {
+        name: "Lisa Permata",
+        avatar: "https://mahasiswa.dinus.ac.id/images/foto/A/A11/2022/A11.2022.14647.jpg",
+        major: "Teknik Informatika",
+        university: "ITS",
+        rating: 4.6,
+        projects: 10,
+        connections: 120,
+      },
+      skills: ["Unity", "C#", "Game Design", "UI/UX", "Mobile Dev"],
+      duration: "5 bulan",
+      members: "2/4",
+      type: "Tugas Akhir",
+      location: "Surabaya",
+      posted: "4 hari yang lalu",
+      image: "https://cdnpro.eraspace.com/media/wysiwyg/artikel/Tahun_2024/April/Game_Matematika_1.jpg",
+      views: 78,
+      applicants: 11,
+      deadline: "25 Feb 2024",
+    },
+    {
+      id: 6,
+      title: "Sistem Manajemen Perpustakaan",
+      description:
+        "Digitalisasi sistem perpustakaan dengan fitur peminjaman online, katalog digital, notifikasi otomatis, dan integrasi RFID",
+      recruiter: {
+        name: "Andi Wijaya",
+        avatar: "https://mahasiswa.dinus.ac.id/images/foto/A/A11/2022/A11.2022.14647.jpg",
+        major: "Teknik Komputer",
+        university: "UNHAS",
+        rating: 4.5,
+        projects: 7,
+        connections: 95,
+      },
+      skills: ["PHP", "CodeIgniter", "MySQL", "Bootstrap", "RFID"],
+      duration: "3 bulan",
+      members: "1/4",
+      type: "Kuliah",
+      location: "Makassar",
+      posted: "5 hari yang lalu",
+      image: "https://i.ytimg.com/vi/qrlD3zlK418/maxresdefault.jpg",
+      views: 56,
+      applicants: 9,
+      deadline: "18 Feb 2024",
+    },
+  ]
+  
+  export const skillOptions = [
+    "React", "Vue.js", "Angular", "Node.js", "Python", "Java", "PHP",
+    "Laravel", "Design", "UI/UX", "Figma", "Adobe Creative", "Video Editing",
+    "Marketing", "Machine Learning", "Data Science", "IoT", "Mobile Dev",
+    "Unity", "Game Design",
+  ]
+  
+  export const locationOptions = ["Jakarta", "Bandung", "Yogyakarta", "Surabaya", "Makassar", "Medan", "Semarang"]
